utilities/requirements.txt
psycopg2-binary>=2.9.0
basketball-reference-web-scraper>=0.6.0
requests>=2.28.0
urllib3>=1.26.0